feat: Picking Images
Add the action method for the camera button and the action method for the album selection.
The UIImagePickerControllerDelegate protocol defines methods that your delegate object must implement to interact with the image picker interface. The methods of this protocol notify your delegate when the user either picks an image or cancels the picker operation.
To be a delegate of the UIImagePickerController, the View Controller class needs to conform to the UINavigationControllerDelegate protocol.
class ViewController: UIViewController, UIImagePickerControllerDelegate,
UINavigationControllerDelegate {
}